+++ kafka/src/test/scala/com/twitter/finatra/kafka/test/integration/FinagleKafkaConsumerIntegrationTest.scala
@@ -1,17 +1,48 @@
 package com.twitter.finatra.kafka.test.integration
 
+import com.twitter.finagle.Init
+import com.twitter.finagle.context.Contexts
+import com.twitter.finagle.filter.PayloadSizeFilter.{ClientRepTraceKey, ClientReqTraceKey}
+import com.twitter.finagle.toggle.flag
+import com.twitter.finagle.tracing.Annotation.BinaryAnnotation
+import com.twitter.finagle.tracing.{Annotation, Flags, NullTracer, Record, Trace, TraceId}
+import com.twitter.finatra.kafka.consumers.TracingKafkaConsumer.{
+  ConsumerGroupIdAnnotation,
+  ConsumerPollAnnotation,
+  ConsumerTopicAnnotation,
+  ProducerTraceIdAnnotation,
+  ClientIdAnnotation => ConsumerClientIdAnnotation
+}
 import com.twitter.finatra.kafka.consumers.{FinagleKafkaConsumer, FinagleKafkaConsumerBuilder}
 import com.twitter.finatra.kafka.domain.{AckMode, KafkaGroupId}
-import com.twitter.finatra.kafka.producers.{FinagleKafkaProducer, FinagleKafkaProducerBuilder}
+import com.twitter.finatra.kafka.producers.TracingKafkaProducer.{
+  ProducerSendAnnotation,
+  ProducerTopicAnnotation,
+  ClientIdAnnotation => ProducerClientIdAnnotation
+}
+import com.twitter.finatra.kafka.producers.{
+  FinagleKafkaProducer,
+  FinagleKafkaProducerBuilder,
+  TracingKafkaProducer
+}
 import com.twitter.finatra.kafka.test.EmbeddedKafka
 import com.twitter.util.{Duration, Time}
-import org.apache.kafka.clients.consumer.KafkaConsumer
+import java.util.concurrent.TimeUnit
+import org.apache.kafka.clients.consumer.{KafkaConsumer, OffsetResetStrategy}
+import org.apache.kafka.clients.producer.{KafkaProducer, ProducerRecord}
 import org.apache.kafka.common.TopicPartition
+import org.apache.kafka.common.header.internals.RecordHeader
 import org.apache.kafka.common.serialization.Serdes
+import org.mockito.ArgumentCaptor
+import org.mockito.Matchers.any
+import org.mockito.Mockito.{spy, times, verify, when}
+import org.mockito.invocation.InvocationOnMock
+import org.mockito.stubbing.Answer
 import scala.collection.JavaConverters._
 
 class FinagleKafkaConsumerIntegrationTest extends EmbeddedKafka {
   private val testTopic = kafkaTopic(Serdes.String, Serdes.String, "test-topic")
+  private val tracingTopic = kafkaTopic(Serdes.String, Serdes.String, "tracing-topic")
   private val emptyTestTopic = kafkaTopic(Serdes.String, Serdes.String, "empty-test-topic")
 
   private def buildTestConsumer(): FinagleKafkaConsumer[String, String] = {
@@ -35,7 +66,19 @@ class FinagleKafkaConsumerIntegrationTest extends EmbeddedKafka {
       .build()
   }
 
-  private def buildNativeTestConsumer(): KafkaConsumer[String, String] = {
+  private def buildNativeTestProducer(): KafkaProducer[String, String] = {
+    FinagleKafkaProducerBuilder()
+      .dest(brokers.map(_.brokerList()).mkString(","))
+      .clientId("test-producer")
+      .ackMode(AckMode.ALL)
+      .keySerializer(Serdes.String.serializer)
+      .valueSerializer(Serdes.String.serializer)
+      .buildClient()
+  }
+
+  private def buildNativeTestConsumer(
+    offsetResetStrategy: OffsetResetStrategy = OffsetResetStrategy.NONE
+  ): KafkaConsumer[String, String] = {
     FinagleKafkaConsumerBuilder()
       .dest(brokers.map(_.brokerList()).mkString(","))
       .clientId("test-consumer")
@@ -43,8 +86,41 @@ class FinagleKafkaConsumerIntegrationTest extends EmbeddedKafka {
       .keyDeserializer(Serdes.String.deserializer)
       .valueDeserializer(Serdes.String.deserializer)
       .requestTimeout(Duration.fromSeconds(1))
+      .autoOffsetReset(offsetResetStrategy)
       .buildClient()
   }
+
+  private def expectedProducerAnnotations: Seq[Annotation] = Seq(
+    BinaryAnnotation("clnt/finagle.label", "kafka.producer"),
+    BinaryAnnotation("clnt/finagle.version", Init.finagleVersion),
+    BinaryAnnotation("clnt/namer.path", brokers.map(_.brokerList()).mkString(",")),
+    BinaryAnnotation(ClientReqTraceKey, 6),
+    BinaryAnnotation(ProducerTopicAnnotation, tracingTopic.topic),
+    BinaryAnnotation(ProducerClientIdAnnotation, "test-producer"),
+    Annotation.ServiceName("kafka.producer"),
+    Annotation.Message(ProducerSendAnnotation),
+    Annotation.ClientSend,
+    Annotation.WireSend,
+    Annotation.ClientRecv,
+    Annotation.WireRecv,
+    Annotation.Rpc("send")
+  )
+
+  private def expectedConsumerAnnotations(traceId: TraceId): Seq[Annotation] = Seq(
+    BinaryAnnotation("clnt/finagle.label", "kafka.consumer"),
+    BinaryAnnotation("clnt/finagle.version", Init.finagleVersion),
+    BinaryAnnotation("clnt/namer.path", brokers.map(_.brokerList()).mkString(",")),
+    BinaryAnnotation(ClientRepTraceKey, 6),
+    BinaryAnnotation(ConsumerTopicAnnotation, tracingTopic.topic),
+    BinaryAnnotation(ConsumerClientIdAnnotation, "test-consumer"),
+    BinaryAnnotation(ConsumerGroupIdAnnotation, "test-group"),
+    BinaryAnnotation(ProducerTraceIdAnnotation, traceId.toString),
+    Annotation.ServiceName("kafka.consumer"),
+    Annotation.Message(ConsumerPollAnnotation),
+    Annotation.ClientRecv,
+    Annotation.WireRecv,
+    Annotation.Rpc("poll")
+  )
 
   test("endOffset returns 0 for empty topic with no events") {
     val consumer = buildTestConsumer()
@@ -145,4 +221,131 @@ class FinagleKafkaConsumerIntegrationTest extends EmbeddedKafka {
     await(consumer.close(Time.now + Duration.fromSeconds(10)))
   }
 
+  test(
+    "consumer should record topic and producer information in the trace when actively tracing and producer trace id is debug") {
+    val producerTraceId = {
+      val id = Trace.nextId
+      id.copy(flags = id.flags.setDebug)
+    }
+    testConsumerTracing(producerTraceId)
+  }
+
+  test(
+    "consumer should record topic and producer information in the trace when actively tracing and producer trace id is sampled") {
+    val producerTraceId = {
+      val id = Trace.nextId
+      id.copy(flags = id.flags.setFlags(Seq(Flags.SamplingKnown, Flags.Sampled)))
+    }
+    testConsumerTracing(producerTraceId)
+  }
+
+  test(
+    "consumer should not record topic and producer information in the trace when not actively tracing") {
+    val producerTraceId = Trace.nextId
+    val nullTracer = spy(new NullTracer())
+
+    flag.overrides.let(com.twitter.finatra.kafka.TracingEnabledToggleId, 1.0) {
+      Trace.letTracer(nullTracer) {
+        val producer = buildNativeTestProducer()
+        val consumer = buildNativeTestConsumer(OffsetResetStrategy.EARLIEST)
+        consumer.subscribe(Seq(tracingTopic.topic).asJava)
+        try {
+          Contexts.local.let(TracingKafkaProducer.TestTraceIdKey, producerTraceId) {
+            producer
+              .send(new ProducerRecord(tracingTopic.topic, null, "Foo", "Bar"))
+              .get(5, TimeUnit.SECONDS)
+          }
+          val records = consumer.poll(java.time.Duration.ofSeconds(5))
+          records.asScala.flatMap(_.headers().asScala) should contain noElementsOf Seq(
+            new RecordHeader(
+              TracingKafkaProducer.TraceIdHeader,
+              TraceId.serialize(producerTraceId.copy(_sampled = Some(false)))))
+        } finally {
+          producer.close(java.time.Duration.ofSeconds(5))
+          consumer.close(java.time.Duration.ofSeconds(5))
+        }
+      }
+    }
+
+    verify(nullTracer, times(0)).record(any[Record])
+  }
+
+  test(
+    "consumer should not record topic and producer information in the trace when tracingEnabledToggle is 0.0") {
+    val producerTraceId = Trace.nextId
+    val nullTracer = spy(new NullTracer())
+    when(nullTracer.isActivelyTracing(any[TraceId])).thenReturn(true)
+
+    flag.overrides.let(com.twitter.finatra.kafka.TracingEnabledToggleId, 0.0) {
+      Trace.letTracer(nullTracer) {
+        val producer = buildNativeTestProducer()
+        val consumer = buildNativeTestConsumer(OffsetResetStrategy.EARLIEST)
+        consumer.subscribe(Seq(tracingTopic.topic).asJava)
+        try {
+          Contexts.local.let(TracingKafkaProducer.TestTraceIdKey, producerTraceId) {
+            producer
+              .send(new ProducerRecord(tracingTopic.topic, null, "Foo", "Bar"))
+              .get(5, TimeUnit.SECONDS)
+          }
+          val records = consumer.poll(java.time.Duration.ofSeconds(5))
+          records.asScala.flatMap(_.headers().asScala) should contain noElementsOf Seq(
+            new RecordHeader(
+              TracingKafkaProducer.TraceIdHeader,
+              TraceId.serialize(producerTraceId.copy(_sampled = Some(false)))))
+        } finally {
+          producer.close(java.time.Duration.ofSeconds(5))
+          consumer.close(java.time.Duration.ofSeconds(5))
+        }
+      }
+    }
+
+    verify(nullTracer, times(0)).record(any[Record])
+  }
+
+  private def testConsumerTracing(producerTraceId: TraceId): Unit = {
+    val traceRecordsCaptor = ArgumentCaptor.forClass(classOf[Record])
+    val nullTracer = spy(new NullTracer())
+
+    flag.overrides.let(com.twitter.finatra.kafka.TracingEnabledToggleId, 1.0) {
+      Trace.letTracer(nullTracer) {
+        val producer = buildNativeTestProducer()
+        val consumer = buildNativeTestConsumer(OffsetResetStrategy.EARLIEST)
+        consumer.subscribe(Seq(tracingTopic.topic).asJava)
+        try {
+          // Always trace for producer
+          when(nullTracer.isActivelyTracing(any[TraceId])).thenReturn(true)
+          Contexts.local.let(TracingKafkaProducer.TestTraceIdKey, producerTraceId) {
+            producer
+              .send(new ProducerRecord(tracingTopic.topic, null, "Foo", "Bar"))
+              .get(5, TimeUnit.SECONDS)
+          }
+          // Answer based on sampling for consumer
+          when(nullTracer.isActivelyTracing(any[TraceId]))
+            .thenAnswer(new Answer[Boolean] { // kept for sbt compatibility
+              override def answer(invocation: InvocationOnMock): Boolean = {
+                // compatibility with older mockito versions
+                invocation.getArguments()(0).asInstanceOf[TraceId].sampled match {
+                  case Some(sampled) => sampled
+                  case None => false
+                }
+              }
+            })
+          val records = consumer.poll(java.time.Duration.ofSeconds(5))
+          records.asScala.flatMap(_.headers().asScala) should contain oneElementOf Seq(
+            new RecordHeader(
+              TracingKafkaProducer.TraceIdHeader,
+              TraceId.serialize(producerTraceId)))
+        } finally {
+          producer.close(java.time.Duration.ofSeconds(5))
+          consumer.close(java.time.Duration.ofSeconds(5))
+        }
+      }
+    }
+
+    verify(nullTracer, times(26)).record(traceRecordsCaptor.capture())
+    val traceRecords = traceRecordsCaptor.getAllValues.asScala
+    val expectedAnnotations =
+      expectedProducerAnnotations ++ expectedConsumerAnnotations(producerTraceId)
+    traceRecords.map(_.annotation) should contain theSameElementsAs expectedAnnotations
+  }
 }